--- a/lib/trace_location/collector.rb
+++ b/lib/trace_location/collector.rb
@@ -83,8 +83,8 @@ module TraceLocation
       end
 
       def remove_indent(source)
-        indent = source.split("\n").first.match(/\A(\s*).+\z/)[1]
-        source.split("\n").map { |line| line.gsub(/\A#{indent}/, '') }.join("\n")
+        indent = source[/\A(\s*)/, 1]
+        source.gsub(/^#{indent}/, '')
       end
     end
   end