--- a/trace_location.gemspec
+++ b/trace_location.gemspec
@@ -23,7 +23,6 @@ Gem::Specification.new do |s|
   s.require_paths = ['lib']
   s.required_ruby_version = '>= 2.6.0'
 
-  s.add_dependency 'binding_of_caller'
   s.add_dependency 'method_source'
 
   s.add_development_dependency 'bundler'